Move JSI related files to their own projects.
