+++ vnext/Chakra/Chakra.vcxitems
@@ -11,16 +11,12 @@
     </ClCompile>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ProjectCapability Include="SourceItemsFromImports" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="$(MSBuildThisFileDirectory)ChakraExecutor.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)ChakraHelpers.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)ChakraNativeModules.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)ChakraPlatform.cpp" />
-    <ClCompile Include="$(MSBuildThisFileDirectory)ChakraJsiRuntime.cpp"  Condition="'$(OSS_RN)' != 'true'"/>
-    <ClCompile Include="$(MSBuildThisFileDirectory)ChakraJsiRuntime_core.cpp"  Condition="'$(OSS_RN)' != 'true'"/>
-    <ClCompile Include="$(MSBuildThisFileDirectory)ChakraJsiRuntime_edgemode.cpp"  Condition="'$(OSS_RN)' != 'true'"/>
     <ClCompile Include="$(MSBuildThisFileDirectory)ChakraTracing.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)ChakraUtils.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)ChakraValue.cpp" />
@@ -31,11 +27,8 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)ChakraExecutor.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)ChakraHelpers.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)ChakraInstanceArgs.h" />
-    <ClInclude Include="$(MSBuildThisFileDirectory)ChakraJsiRuntimeArgs.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)ChakraNativeModules.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)ChakraPlatform.h" />
-    <ClInclude Include="$(MSBuildThisFileDirectory)ChakraJsiRuntimeFactory.h" />
-    <ClInclude Include="$(MSBuildThisFileDirectory)ChakraJsiRuntime.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)ChakraTracing.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)ChakraUtils.h" />
     <ClInclude Include="$(MSBuildThisFileDirectory)ChakraValue.h" />
