+++ vnext/JSI/Desktop/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="ChakraCore.Debugger" version="0.0.0.39" targetFramework="native" />
+  <package id="Microsoft.ChakraCore.vc140" version="1.11.10" targetFramework="native" developmentDependency="true" />
+</packages>
