+++ vnext/JSI/Shared/ChakraJsiRuntime.cpp
@@ -1,8 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#include "pch.h"
-
 #include "ChakraJsiRuntime.h"
 #include "ChakraJsiRuntimeArgs.h"
